+++ test/lib/sstable_utils.hh
@@ -233,6 +233,10 @@ public:
     void reload_reclaimed_components() {
         _sst->reload_reclaimed_components().get();
     }
+
+    const utils::filter_ptr& get_filter() const {
+        return _sst->_components->filter;
+    }
 };
 
 inline auto replacer_fn_no_op() {
