+++ utils/i_filter.cc
@@ -32,6 +32,17 @@ filter_ptr i_filter::get_filter(int64_t num_elements, double max_false_pos_proba
     return filter::create_filter(spec.K, num_elements, spec.buckets_per_element, fformat);
 }
 
+size_t i_filter::get_filter_size(int64_t num_elements, double max_false_pos_probability) {
+    if (max_false_pos_probability >= 1.0) {
+        return 0;
+    }
+
+    int buckets_per_element = bloom_calculations::max_buckets_per_element(num_elements);
+    auto spec = bloom_calculations::compute_bloom_spec(buckets_per_element, max_false_pos_probability);
+
+    return filter::get_bitset_size(num_elements, spec.buckets_per_element) / 8;
+}
+
 hashed_key make_hashed_key(bytes_view b) {
     std::array<uint64_t, 2> h;
     utils::murmur_hash::hash3_x64_128(b, 0, h);
