+++ utils/i_filter.hh
@@ -52,5 +52,10 @@ struct i_filter {
      *         filter.
      */
     static filter_ptr get_filter(int64_t num_elements, double max_false_pos_prob, filter_format format);
+
+    /**
+     * @return the size of the smallest filter (in bytes), according to the conditions described at get_filter()
+     */
+    static size_t get_filter_size(int64_t num_elements, double max_false_pos_prob);
 };
 }
